Created AnimalVIewController

+++ src/main/java/org/example/zoomanagementsystem/Model/Lion.java
@@ -1,7 +1,7 @@
 package org.example.zoomanagementsystem.Model;
 
 public class Lion extends Animal {
-    public Lion(String pName, int pAge) {
+    public Lion(String pName, double pAge) {
         super(pName, pAge);
     }
 }
